Start adding test and possibly some fixes for #481
Conversion to GFF3Feature includes source and score and sets ID and
Parent attributes. However, export to GFF from UI fails.
