+++ packages/apollo-shared/src/GFF3/index.ts
@@ -1,2 +1,3 @@
+export * from './annotationFeatureToGFF3'
 export * from './gffReservedKeys'
 export * from './gff3ToAnnotationFeature'
